fix: lint errors, change file name
